Revise repo structure
First attempt to parse through array processing tools and organize them
into modules
