+++ array_processing/algorithms/__init__.py
@@ -1,1 +1,2 @@
 from .array_algorithms import wlsqva, fk_freq
+from .helpers import wlsqva_proc, array_plot, getrij
